Make dockered_chunked_call self contained and add entrypoint
