+++ setup.py
@@ -16,7 +16,8 @@ kwargs = dict(
     package_dir={'': 'src'},
     packages=find_packages('src'),
     entry_points={
-        'console_scripts': ['toil-vg = toil_vg.vg_evaluation_pipeline:main']})
+        'console_scripts': ['toil-vg = toil_vg.vg_evaluation_pipeline:main',
+                            'dockered_chunked_call = toil_vg.dockered_chunked_call:main']})
 
 class PyTest(TestCommand):
     user_options = [('pytest-args=', 'a', "Arguments to pass to py.test")]
