+++ src/toil_vg/vg_evaluation_pipeline.py
@@ -567,85 +567,6 @@ def run_merge_vcf(job, options, index_dir_id, vcf_file_key_list):
 
     return downloadList
 
-
-def run_calling(job, options, index_dir_id, alignment_file_key, path_name, path_size):
-    
-    RealTimeLogger.get().info("Running variant calling on path {} from alignment file {}".format(path_name, alignment_file_key))
-    
-    # Set up the IO stores each time, since we can't unpickle them on Azure for
-    # some reason.
-    input_store = IOStore.get(options.input_store)
-    out_store = IOStore.get(options.out_store)
-    
-    # Define work directory for docker calls
-    work_dir = job.fileStore.getLocalTempDir()
-    
-    # Download the indexed graph to a directory we can use
-    graph_dir = work_dir
-    read_global_directory(job.fileStore, index_dir_id, graph_dir)
-
-    # How long did the alignment take to run, in seconds?
-    run_time = None
-
-    # We know what the xg file in there will be named
-    xg_file = "{}/graph.vg.xg".format(graph_dir)
-
-    # Download the alignment
-    alignment_file = "{}/{}.gam".format(work_dir, options.sample_name)
-    out_store.read_input_file(alignment_file_key, alignment_file)
-    
-    variant_call_dir = work_dir
-
-    # run chunked_call
-    chunks = dockered_chunked_call(job, options, out_store, work_dir, index_dir_id, job.cores, xg_file, alignment_file, path_name, path_size, options.sample_name, variant_call_dir, options.call_chunk_size, options.overlap, options.filter_opts, options.pileup_opts, options.call_opts, options.overwrite)
-
-    vcf_file_key = job.addFollowOnJobFn(merge_vcf_chunks, options, index_dir_id, path_name, path_size, chunks, options.overwrite, cores=2, memory="4G", disk="2G").rv()
- 
-    RealTimeLogger.get().info("Completed variant calling on path {} from alignment file {}".format(path_name, alignment_file_key))
-
-    return vcf_file_key
-
-def dockered_chunked_call(job, options, out_store, work_dir, index_dir_id, threads, xg_path, gam_path, path_name, path_size, sample_name, out_dir, chunk=10000000,
-                          overlap=2000, filter_opts="-r 0.9 -d 0.05 -e 0.05 -afu -s 1000 -o 10",
-                          pileup_opts="-w 40 -m 10 -q 10", call_opts="-b 0.4 -f 0.25 -d 10 -s 1",
-                          overwrite=True):
-    """
-    dockered_chunked_call IS NOT A TOIL JOB FUNCTION. It is just a helper function.
-    """
-    
-    #if not os.path.isdir(out_dir):
-    #   os.makedirs(out_dir)
-
-    # make things slightly simpler as we split overlap
-    # between adjacent chunks
-    assert overlap % 2 == 0
-
-    # compute overlapping chunks
-    chunks = make_chunks(path_name, path_size, chunk, overlap)
-    
-    # split the gam in one go
-    chunk_gam(options.drunner, gam_path, xg_path, path_name, work_dir,
-              chunks, filter_opts, overwrite)
-
-    # call every chunk in series
-    for chunk_i, chunk in enumerate(chunks):
-        # make the graph chunk
-        chunk_vg(options.drunner, xg_path, path_name, work_dir, chunks, chunk_i, overwrite)
-        
-        vg_path = chunk_base_name(path_name, work_dir, chunk_i, ".vg")
-        gam_path = chunk_base_name(path_name, work_dir, chunk_i, ".gam")
-        
-        # upload split gam files
-        out_store.write_output_file(vg_path, os.path.basename(vg_path))
-        out_store.write_output_file(gam_path, os.path.basename(gam_path))
-        
-        job.addChildJobFn(call_chunk, options, index_dir_id, xg_path, path_name, chunks, chunk_i,
-                           path_size, overlap,
-                           pileup_opts, call_opts,
-                           sample_name, threads,
-                           overwrite, cores="{}".format(threads), memory="4G", disk="2G")
-    return chunks
-
 def run_upload(job, options, uploadList):
     """
     Upload and file in uploadList to the remote IO store specified
